Refactored google elevator service to hook.
Applied to both maps.
